Test for inundation.

+++ glofris_postprocess_edwin_modified.py
@@ -784,7 +784,7 @@ def get_return_period_gumbel(p_zero_in_pcraster, loc_in_pcraster, scale_in_pcras
     #~ os.system(cmd)
 
     # transform the reduced variate into a probability (residual after removing the zero volume probability)
-    p_residual = np.minimum(np.maximum(np.exp(-np.exp(-np.longdouble(reduced_variate))), 0.0), 1.0)
+    p_residual = np.minimum(np.maximum(np.exp(-np.exp(-np.longdouble(reduced_variate))), np.longdouble(0.0)), np.longdouble(1.0))
 
     # tranform from non-zero only distribution to zero-included distribution
     p = np.minimum(np.maximum(p_residual*(1.0 - p_zero) + p_zero, p_zero), max_p)  # never larger than max_p #
